Refactoring, create cmd  and internal
