+++ Dockerfile
@@ -1,19 +1,19 @@
 ARG GO_VERSION=1.14
 FROM registry.access.redhat.com/ubi8/go-toolset:latest AS builder
-WORKDIR /scheduler/
+WORKDIR /agnostics/
 
 USER root
 RUN chown -R ${USER_UID}:0 /scheduler
 USER ${USER_UID}
 
-COPY scheduler/ ./
+COPY ./* ./
 ENV GOOS=linux
-RUN go build
+RUN go build ./cmd/scheduler
 
 FROM registry.access.redhat.com/ubi8/ubi-minimal:latest AS deploy
-WORKDIR /scheduler/
+WORKDIR /agnostics/
 USER ${USER_UID}
-COPY github_known_hosts /ssh/known_hosts
+COPY build/github_known_hosts /ssh/known_hosts
 env SSH_KNOWN_HOSTS /ssh/known_hosts
-COPY --from=builder /scheduler/scheduler ./
+COPY --from=builder /agnostics/scheduler ./
 CMD ["./scheduler"]
