+++ deployments/helm/templates/deployment-redis.yml
@@ -1,0 +1,89 @@
+apiVersion: extensions/v1beta1
+kind: Deployment
+metadata:
+  annotations:
+    deployment.kubernetes.io/revision: "2"
+    template.alpha.openshift.io/wait-for-ready: "true"
+  labels:
+    app: scheduler
+    template: deploy-scheduler-template
+  name: scheduler-redis
+  namespace: scheduler-gpe-demo
+  resourceVersion: "70023207"
+  selfLink: /apis/extensions/v1beta1/namespaces/scheduler-gpe-demo/deployments/scheduler-redis
+  uid: f521304a-b243-11ea-b0aa-0a54efb2fbb6
+spec:
+  progressDeadlineSeconds: 600
+  replicas: 1
+  revisionHistoryLimit: 10
+  selector:
+    matchLabels:
+      name: scheduler-redis
+  strategy:
+    type: Recreate
+  template:
+    metadata:
+      creationTimestamp: null
+      labels:
+        name: scheduler-redis
+    spec:
+      containers:
+      - env:
+        - name: REDIS_PASSWORD
+          valueFrom:
+            secretKeyRef:
+              key: database-password
+              name: scheduler-redis
+        image: redislabs/rejson:latest
+        imagePullPolicy: Always
+        livenessProbe:
+          failureThreshold: 3
+          initialDelaySeconds: 30
+          periodSeconds: 10
+          successThreshold: 1
+          tcpSocket:
+            port: 6379
+          timeoutSeconds: 1
+        name: redis
+        ports:
+        - containerPort: 6379
+          protocol: TCP
+        resources:
+          limits:
+            memory: 512Mi
+        securityContext:
+          capabilities: {}
+          privileged: false
+        terminationMessagePath: /dev/termination-log
+        terminationMessagePolicy: File
+        volumeMounts:
+        - mountPath: /var/lib/redis/data
+          name: scheduler-redis-data
+      dnsPolicy: ClusterFirst
+      restartPolicy: Always
+      schedulerName: default-scheduler
+      securityContext: {}
+      terminationGracePeriodSeconds: 30
+      volumes:
+      - name: scheduler-redis-data
+        persistentVolumeClaim:
+          claimName: scheduler-redis
+status:
+  availableReplicas: 1
+  conditions:
+  - lastTransitionTime: "2020-06-24T08:09:39Z"
+    lastUpdateTime: "2020-06-24T08:09:39Z"
+    message: Deployment has minimum availability.
+    reason: MinimumReplicasAvailable
+    status: "True"
+    type: Available
+  - lastTransitionTime: "2020-06-19T15:45:53Z"
+    lastUpdateTime: "2020-06-24T08:09:39Z"
+    message: ReplicaSet "scheduler-redis-5ddd95d55f" has successfully progressed.
+    reason: NewReplicaSetAvailable
+    status: "True"
+    type: Progressing
+  observedGeneration: 2
+  readyReplicas: 1
+  replicas: 1
+  updatedReplicas: 1
