+++ deployments/openshift/deploy-scheduler-template.yaml
