+++ go.mod
@@ -6,9 +6,9 @@ require (
 	github.com/go-git/go-git/v5 v5.1.0
 	github.com/gomodule/redigo v1.8.2
 	github.com/julienschmidt/httprouter v1.3.0
-	golang.org/x/crypto v0.0.0-20200604202706-70a84ac30bf9
+	golang.org/x/crypto v0.0.0-20200604202706-70a84ac30bf9 // indirect
 	golang.org/x/net v0.0.0-20200602114024-627f9648deb9 // indirect
 	golang.org/x/sys v0.0.0-20200610111108-226ff32320da // indirect
-	gopkg.in/src-d/go-git.v4 v4.13.1
+	gopkg.in/src-d/go-git.v4 v4.13.1 // indirect
 	gopkg.in/yaml.v2 v2.3.0
 )
