+++ internal/modules/taints_test.go
@@ -2,7 +2,7 @@ package modules
 
 import (
 	"testing"
-	"github.com/redhat-gpe/agnostics/api/v1"
+	"github.com/redhat-gpe/agnostics/internal/api/v1"
 	"sort"
 )
